Added Lucene to the project
The dependency has been added in maven and a goal that copies
dependencies to target/lib has been added such that we can
reference the libraries in the Eclipse classpath and the
manifest of the OSGI bundles.

I added a very basic integration test that shows that the
classpath is set up properly and that Lucene is working.

Fixes #5