Stores occurrences of methods in Lucene
The plugin now keeps an Index (using Lucene) of all the places
methods are mentioned in the code-base.

The initial index will be created upon start-up if no index
exists already. It will index the entire workspace and traverse
the parse-tree of all Scala sources.

Using a background job we continuously update the index of all
files that are changed so it always reflect the state of the files
in the workspace. We currently assume that no other applications
will change the source files when Eclipse is closed.

Also added a test that shows that we can store and retrieve information
from Lucene and that it's able to serialize and de-serialize the
abstraction we use to represent an Occurrence.

Fixes #6