Provides API for finding potential occurrences
This includes a new component, ProjectFinder, which is
responsible for finding the transitive closure of project
dependencies in the Eclipse workspace.

Additionally I added a new method to Index that allows the
API user to search for a given word in the index of a set
of projects.

Tests are included.

Fixes #1001663