Provides API for finding exact occurrences of enity at location
This ties together the Index and SearchPresentationCompiler to
find exact occurrences of a method at a given location.

Tests have been provided that tests that these components play
together nicely but I've kept them simple as the majority of the
tests are written for each of the components (Index, SPC).

Fixes #1001685