Can now find occurrences of apply
The Finder could previously not find these occurrences as
the don't appear Apply(Select(apply),...) but rather
Apply(Ident(Foo),..).

I fixed it by making it possible to search for multiple
words when looking for occurrences (this will also be needed
later when searching for aliases). When searching for an apply
invocation i will now search for explicit invocations (i.e. apply)
and implicit invocations (i.e. Foo). So it will search for apply and
the name of the owner of the apply method.

Fixes #1001736