Fixes race condition in importers
This commit fixes a race condition that would appear when using
the symbol importers.

The problem was that we were importing a symbol from one presentation
compiler to another, so it would be accessing thead un-safe datastructures
in both presentaion compilers when importing (one for reading the symbol
fields, the other for writing them).

To fix this we make sure that the symbol (and it's owner chain) is
fully initialized in the PC it originates from before trying to import
it into another PC.

Wasn't able to reproduce this in tests.

Fixes #39 (Github)