Fix exception when showing hits for closed project/deleted files
A Java Model Exception would be thrown whenever the search view
was showing results from files that no longer existed or files
that belonged to projects that were no longer open.

This also fixes a less serious bug. When performing a search Eclipse
will now corretly show markers in the file that is open if any
results are contained in that file. Before you had to open another
file and return to the original one for the markers to show up.

Fixes #1001741