Fixing the OccurrenceCollectorTest for Scala 2.11.x
There is a slight difference between Scala 2.10.x and 2.11.x.

In Scala 2.11.x calls that are placed inside of constructors
that are generated by the compiler have NoPosition as their pos.
(It's a bit weird that they had a position in Scala 2.10.x
anyway since they are synthetic.)

This takes care of this edge case by being more specific about
what kind of occurrence we expect in the test; this should be
reasonbale since we don’t know how to deal with
synthetic symbols yet anyway.