Remove 'contextId' from type hierarchy binding
The 'contextId' is used to restrict the "scope" of a binding. In
this case I used it to make sure that our type hierarchy could
only be invoked from inside a scala editor. But someting whent
fishy and the key binding would only work for the first invocation

I have a feeling this might have something to do with the fact that
the Scala editor isn't a real scala editor yet but a little
aspect hack.