Workaround for race condition in symbol importer
Looks like the symbol importer module isn't thread-safe. In particular, before
importing a symbol from one presentation compiler to another, we need to make
sure both the symbol and its owner chain are **fully initialized**, and the
symbol's ``sourceModel`` is also initialized. It's brittle, we know it, but
this seems to ensure we don't initialize symbols in the wrong presentation
compiler thread.

Fixes #1001826